Final version of my project.
I coded an archery board which fills in the colour of the rings if the mouse is moved over it. If the mouse is clicked outside of the board the whole board is displayed. I added an image of a forest as the background. I practiced incorporating classes and methods into my code with this project.